requirements.txt
alpaca-py
python-dotenv
xai-sdk
pytz
tavily-python
